+++ compare_with_dcp.py
@@ -6,7 +6,7 @@ import select
 import json
 
 from helper_files.constants.tier1_mapping import entity_types, all_entities
-from helper_files.file_io import open_spreadsheet, get_label
+from helper_files.utils import open_spreadsheet, get_label
 
 # Open cellxgene spreadsheet
 # Open DCP spreadsheet
